- restructuring components. - moving to module .css
