+++ src/components/base/DataTable.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import styles from '../assets/DataTable.module.css';
+import styles from '../../assets/DataTable.module.css';
 
 const DataTable = ({ name, data }) => {
   const [sortConfig, setSortConfig] = useState({ key: null, direction: 'asc' });
