+++ src/components/container/NavigationBar.js
@@ -1,23 +1,23 @@
 import React, { useContext } from 'react';
 import { Link } from 'react-router-dom';
-import UserContext from '../context/user_context';
-import SignOut from './sign_out';
-import '../assets/navigation_bar.css'; 
+import UserContext from '../../context/user_context';
+import SignOut from '../sign_out';
+import styles from '../../assets/navigation-bar.module.css'; 
 
 const NavigationBar = () => {
   const { isLoggedIn, user } = useContext(UserContext);
   const initials = isLoggedIn && user ? `${user.first_name[0]}${user.last_name[0]}` : '';
 
   return (
-    <header className="header">
-      <div className="header-wrapper">
-        <div className="logo">
-            <Link to="/simulate" className="header-link">
+    <header className={styles.header}>
+      <div className={styles.header_wrapper}>
+        <div className={styles.logo}>
+            <Link to="/simulate">
                 <h1>Fyn-Front</h1>
             </Link>
             <p>Creativity Leads Innovation</p>
         </div>
-        <nav className="navbar">
+        <nav className={styles.navbar}>
             {isLoggedIn ? (
               <>
               <Link to="/simulate">Simulate</Link>
